// src/app/models/product.model.ts
export class ProductModel {
  ABC?: number;
  code?: string;
  pack?: number;
  box?: number;
  weight?: number;
  price?: number;
  category?: string;
  group?: string;
  name?: string;
  discription?: string;
  pass?: string;
  tech?: string;
  img?: string;
  sae?: string;
  fullDiscription?: string;
  using?: string | null;
  benefits?: string | null;
  recomendation?: string | null;
  instructions?: string | null;
  quantity?: number | null;
  approvs?: string[];
  approvsEqual?: string[];
  approvsRecomendation?: string[];
}